+++ l10n_br_account_product/models/res_company.py
@@ -79,19 +79,3 @@ class ResCompany(models.Model):
         'account.tax', string='Other Costs Sale Tax',
         domain=[('domain', '=', 'other_costs')])
     accountant_cnpj_cpf = fields.Char(size=18, string='CNPJ/CPF Contador')
-
-
-class L10nBrTaxDefinitionCompanyProduct(L10nBrTaxDefinition, models.Model):
-    _name = 'l10n_br_tax.definition.company.product'
-
-    company_id = fields.Many2one('res.company', 'Empresa')
-    tax_ipi_guideline_id = fields.Many2one(
-        'l10n_br_account_product.ipi_guideline', string=u'Enquadramento IPI')
-    tax_icms_relief_id = fields.Many2one(
-        'l10n_br_account_product.icms_relief', string=u'Desoneração ICMS')
-
-    _sql_constraints = [
-        ('l10n_br_tax_definition_tax_id_uniq',
-         'unique (tax_id, company_id)',
-         u'Imposto já existente nesta empresa!')
-    ]
